+++ CHANGES.md
@@ -1,8 +1,12 @@
+# 2019-12-23
+
+- added `voting_period` param to `activate` ACTION (must be current or next)
+- add `auto_complete` which pushes `complete` deferred transaction using `next_voting_period` as the delay
+
 # 2019-12-22
 
 - add `check_completed` check, cannot perform `activate`/`vote` ACTIONS during completed voting period phase
 - add check for "cannot exceed 100 proposals per single voting period"
-- added `voting_period` param to `activate` ACTION (must be current or next)
 
 # 2019-12-21